CONFLICT (content): Merge conflict in llvm/test/Transforms/SimplifyCFG/invoke_unwind.ll
  CONFLICT (content): Merge conflict in llvm/lib/Transforms/Scalar/JumpThreading.cpp

  commit 10f2a0d662d8d72eaac48d3e9b31ca8dc90df5a4
  Author: Arthur Eubanks <aeubanks@google.com>
  Date:   Wed Sep 30 12:11:46 2020 -0700

      Use uint64_t for branch weights instead of uint32_t

      CallInst::updateProfWeight() creates branch_weights with i64 instead of i32.
      To be more consistent everywhere and remove lots of casts from uint64_t
      to uint32_t, use i64 for branch_weights.

      Reviewed By: davidxl

      Differential Revision: https://reviews.llvm.org/D88609

+++ llvm/lib/Transforms/Instrumentation/PGOInstrumentation.cpp
@@ -117,6 +117,7 @@
 #include <cstdint>
 #include <memory>
 #include <numeric>
+#include <stdint.h>
 #include <string>
 #include <unordered_map>
 #include <utility>
@@ -1868,7 +1869,7 @@ void llvm::setProfMetadata(Module *M, Instruction *TI,
   MDBuilder MDB(M->getContext());
   assert(MaxCount > 0 && "Bad max count");
   uint64_t Scale = calculateCountScale(MaxCount);
-  SmallVector<unsigned, 4> Weights;
+  SmallVector<uint64_t, 4> Weights;
   for (const auto &ECI : EdgeCounts)
     Weights.push_back(scaleBranchCount(ECI, Scale));
